odin backend: keyword→enum variants, Int→int, struct literal fixes, str→fmt.tprintf
Emitter: keyword symbols (:air) emit as .Air enum variants in all
expression positions (not just condp tests). Int maps to Odin's
native int (not i64) for natural array indexing. Struct literal
field names and enum definitions drop erroneous dot prefixes.
str function uses fmt.tprintf instead of requiring beagle_rt.

Checker: defenum registers enum type names; Keyword type is
compatible with registered enum types (keyword literals accepted
as enum values at type-check time).

+++ beagle-test/tests/fixtures/odin-golden/01-record.odin
@@ -3,6 +3,6 @@ package g
 
 
 Point :: struct {
-    x: i64,
-    y: i64,
+    x: int,
+    y: int,
 }
